moved stuff around, first framed screens ready for wire and polish
