+++ src/components/Headers/Header.js
@@ -7,9 +7,9 @@ import {
 import React, { Component } from "react";
 import Icon from 'react-native-vector-icons/FontAwesome';
 import styles from "./HeaderStyles";
-import { stackNavigator } from "react-navigation";
+import { stackNavigator, withNavigation } from "react-navigation";
 import ColorConstants from "../../assets/ColorConstants";
-import ComponentTest from "../../screens/ComponentTest";
+import ComponentTest from "../../screens/FramedScreens/ComponentTest";
 const bgImage = require("../../assets/main-bg.png")
